changed the storage of raws for polygons back to 8 bytes as this apparently what ilwis3 expcects

+++ ilwis3connector/ilwis3featureconnector.h
@@ -58,7 +58,7 @@ private:
         output_file.write((char *)&y, 8);
     }
     void writeLine(const geos::geom::LineString *line, std::ofstream &output_file, long raw);
-    void writePolygon(const geos::geom::Polygon *polygon, std::ofstream &output_file, long raw);
+    void writePolygon(const geos::geom::Polygon *polygon, std::ofstream &output_file, double raw);
     void writePoint(const geos::geom::Point *point, std::ofstream &output_file, long raw);
     void addFeatures(map<quint32, vector<geos::geom::Geometry *> > &geometries, FeatureCoverage *fcoverage, ITable &tbl, IlwisTypes tp);
     bool storeBinaryDataTable(IlwisObject *obj, IlwisTypes tp, const QString &baseName);
